# Speedup audb.available()