--- a/setup.cfg
+++ b/setup.cfg
@@ -27,7 +27,7 @@ classifiers =
 [options]
 packages = find:
 install_requires =
-    audbackend >=0.3.7
+    audbackend >=0.3.10
     audeer >=1.14.0
     audformat >=0.11.1,<2.0.0
     audiofile >=1.0.0