Speedup the call to `audb.available()`.

So far we have used `backend.glob()` to get a list of all databases.
As the structure how we store the data is fixed, we can instead simply list the content of the folder.

This improves execution times on our internal repos.
E.g. over the VPN I get for calling `df = audb.available()`:

* Before: 237s
* After: 100s

I repeated the measurement internally on our compute servers and get:

* Before: 8.1s
* After: 3.9s

/cc @ChristianGeng